+++ ThirdPartyAdapters/meta/CHANGELOG.md
@@ -1,7 +1,9 @@
 ## Meta Audience Network Android Mediation Adapter Changelog
 
 #### Next Version
+
 - Updated native ad implementation to use the updated `NativeAdMapper` API.
+- Destroy `NativeAdBase` and `MediaView` when `untrackView` is called to release resources.
 
 #### Version 6.21.0.2
 - Added property to build the adapter with GMA Next-Gen SDK dependency.